LPS-179353 Extract ScreenNavigationEntry methods from CommerceSubscriptionEntryShipmentsScreenNavigationCategory to CommerceSubscriptionEntryShipmentsScreenNavigationEntry and extends CommerceSubscriptionEntryShipmentsScreenNavigationCategory in it
