+++ modules/apps/commerce/commerce-subscription-web/src/main/java/com/liferay/commerce/subscription/web/internal/frontend/taglib/servlet/taglib/CommerceSubscriptionEntryShipmentsScreenNavigationCategory.java
@@ -14,29 +14,11 @@
 
 package com.liferay.commerce.subscription.web.internal.frontend.taglib.servlet.taglib;
 
-import com.liferay.commerce.constants.CommerceConstants;
-import com.liferay.commerce.model.CommerceSubscriptionEntry;
-import com.liferay.commerce.payment.service.CommercePaymentMethodGroupRelLocalService;
-import com.liferay.commerce.product.util.CPSubscriptionTypeJSPContributorRegistry;
-import com.liferay.commerce.product.util.CPSubscriptionTypeRegistry;
-import com.liferay.commerce.service.CommerceOrderItemLocalService;
-import com.liferay.commerce.service.CommerceSubscriptionEntryLocalService;
 import com.liferay.commerce.subscription.web.internal.constants.CommerceSubscriptionEntryScreenNavigationConstants;
-import com.liferay.commerce.subscription.web.internal.display.context.CommerceSubscriptionEntryDisplayContext;
 import com.liferay.frontend.taglib.servlet.taglib.ScreenNavigationCategory;
-import com.liferay.frontend.taglib.servlet.taglib.ScreenNavigationEntry;
-import com.liferay.frontend.taglib.servlet.taglib.util.JSPRenderer;
 import com.liferay.portal.kernel.language.Language;
-import com.liferay.portal.kernel.security.permission.resource.PortletResourcePermission;
-import com.liferay.portal.kernel.util.WebKeys;
-
-import java.io.IOException;
 
 import java.util.Locale;
-
-import javax.servlet.ServletContext;
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
 
 import org.osgi.service.component.annotations.Component;
 import org.osgi.service.component.annotations.Reference;
@@ -45,15 +27,11 @@ import org.osgi.service.component.annotations.Reference;
  * @author Alessio Antonio Rendina
  */
 @Component(
-	property = {
-		"screen.navigation.category.order:Integer=30",
-		"screen.navigation.entry.order:Integer=10"
-	},
-	service = {ScreenNavigationCategory.class, ScreenNavigationEntry.class}
+	property = "screen.navigation.category.order:Integer=30",
+	service = ScreenNavigationCategory.class
 )
 public class CommerceSubscriptionEntryShipmentsScreenNavigationCategory
-	implements ScreenNavigationCategory,
-			   ScreenNavigationEntry<CommerceSubscriptionEntry> {
+	implements ScreenNavigationCategory {
 
 	@Override
 	public String getCategoryKey() {
@@ -62,13 +40,8 @@ public class CommerceSubscriptionEntryShipmentsScreenNavigationCategory
 	}
 
 	@Override
-	public String getEntryKey() {
-		return getCategoryKey();
-	}
-
-	@Override
 	public String getLabel(Locale locale) {
-		return _language.get(locale, getCategoryKey());
+		return language.get(locale, getCategoryKey());
 	}
 
 	@Override
@@ -77,63 +50,7 @@ public class CommerceSubscriptionEntryShipmentsScreenNavigationCategory
 			SCREEN_NAVIGATION_KEY_COMMERCE_SUBSCRIPTION_ENTRY;
 	}
 
-	@Override
-	public void render(
-			HttpServletRequest httpServletRequest,
-			HttpServletResponse httpServletResponse)
-		throws IOException {
-
-		CommerceSubscriptionEntryDisplayContext
-			commerceSubscriptionEntryDisplayContext =
-				new CommerceSubscriptionEntryDisplayContext(
-					_commercePaymentMethodGroupRelLocalService,
-					_commerceSubscriptionEntryLocalService,
-					_commerceOrderItemLocalService,
-					_cpSubscriptionTypeJSPContributorRegistry,
-					_cpSubscriptionTypeRegistry, httpServletRequest,
-					_portletResourcePermission);
-
-		httpServletRequest.setAttribute(
-			WebKeys.PORTLET_DISPLAY_CONTEXT,
-			commerceSubscriptionEntryDisplayContext);
-
-		_jspRenderer.renderJSP(
-			_servletContext, httpServletRequest, httpServletResponse,
-			"/commerce_subscription_entry/shipments.jsp");
-	}
-
 	@Reference
-	private CommerceOrderItemLocalService _commerceOrderItemLocalService;
-
-	@Reference
-	private CommercePaymentMethodGroupRelLocalService
-		_commercePaymentMethodGroupRelLocalService;
-
-	@Reference
-	private CommerceSubscriptionEntryLocalService
-		_commerceSubscriptionEntryLocalService;
-
-	@Reference
-	private CPSubscriptionTypeJSPContributorRegistry
-		_cpSubscriptionTypeJSPContributorRegistry;
-
-	@Reference
-	private CPSubscriptionTypeRegistry _cpSubscriptionTypeRegistry;
-
-	@Reference
-	private JSPRenderer _jspRenderer;
-
-	@Reference
-	private Language _language;
-
-	@Reference(
-		target = "(resource.name=" + CommerceConstants.RESOURCE_NAME_COMMERCE_SUBSCRIPTION + ")"
-	)
-	private PortletResourcePermission _portletResourcePermission;
-
-	@Reference(
-		target = "(osgi.web.symbolicname=com.liferay.commerce.subscription.web)"
-	)
-	private ServletContext _servletContext;
+	protected Language language;
 
 }
